Refactor logger, introduce output
